Align gallery controller types
